Warn user when access token is about to expire (#42)
Update test app authorization code to add a warning notification when
the user access token is about to expire. Additionally, rename
`userAuthorizationClient` to `authorizationClient` and make it
non-nullable. Instead, the client will always be present but it will
check whether the token is valid at the moment `getAccessToken` is
called.