Release packages [publish docs] (#129)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to master, this PR
will be updated.

# Releases
## @itwin/saved-views-client@0.5.2

### Patch Changes

####
[0.5.2](https://github.com/iTwin/saved-views/tree/v0.5.2-client/packages/saved-views-client)
- 2025-05-22

Remove content type header when request body is
empty@itwin/saved-views-client

 ## @itwin/saved-views-react@0.9.8

Co-authored-by: imodeljs-admin <imodeljs-admin@users.noreply.github.com>