Release packages [publish docs] (#133)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to master, this PR
will be updated.

# Releases
## @itwin/saved-views-react@1.0.0

### Major Changes

####
[1.0.0](https://github.com/iTwin/saved-views/tree/v1.0.0-react/packages/saved-views-react)
- 2025-06-03

Change models and categories fallback setting option to seperate and
further define settings for models and
categories@itwin/saved-views-react

Co-authored-by: imodeljs-admin <imodeljs-admin@users.noreply.github.com>