Release packages [publish docs] (#135)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to master, this PR
will be updated.

# Releases
## @itwin/saved-views-react@1.1.0

### Minor Changes

####
[1.1.0](https://github.com/iTwin/saved-views/tree/v1.1.0-react/packages/saved-views-react)
- 2025-06-05

Add subcategories option in applyView settings.@itwin/saved-views-react

Co-authored-by: imodeljs-admin <imodeljs-admin@users.noreply.github.com>