Release packages [publish docs] (#144)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to master, this PR
will be updated.

# Releases
## @itwin/saved-views-client@0.5.3

### Patch Changes

####
[0.5.3](https://github.com/iTwin/saved-views/tree/v0.5.3-client/packages/saved-views-client)
- 2025-08-07

Add support for properties and queryParams in
ImageMapLayerProps@itwin/saved-views-client

 ## @itwin/saved-views-react@1.1.2

### Patch Changes

####
[0.5.3](https://github.com/iTwin/saved-views/tree/v0.5.3-client/packages/saved-views-client)
- 2025-08-07

Add support for properties and queryParams in
ImageMapLayerProps@itwin/saved-views-client

Co-authored-by: imodeljs-admin <imodeljs-admin@users.noreply.github.com>