Release packages [publish docs] (#155)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to master, this PR
will be updated.

# Releases
## @itwin/saved-views-client@0.6.0

### Minor Changes

####
[0.6.0](https://github.com/iTwin/saved-views/tree/v0.6.0-client/packages/saved-views-client)
- 2025-10-07

Expose CommonJS build for
@itwin/saved-views-client@itwin/saved-views-client

 ## @itwin/saved-views-react@1.2.1

Co-authored-by: imodeljs-admin <imodeljs-admin@users.noreply.github.com>